requirements.txt
annotated-types==0.6.0
asgiref==3.7.2
asttokens==2.4.1
colorama==0.4.6
crispy-bootstrap4==2023.1
Django==4.2.10
django-cors-headers==4.3.1
django-crispy-forms==2.1
django-extensions==3.2.3
django-ninja==1.1.0
django-tables2==2.7.0
django-utils
executing==2.0.1
Faker==23.2.1
icecream==2.1.3
pydantic==2.6.1
pydantic-core==2.16.2
pygments==2.17.2
six==1.16.0
sqlparse==0.4.4
typing-extensions==4.9.0
django-q==1.3.9
psycopg2>=2.9.9,<2.9.10
#uWSGI>=2.0.19.1,<2.1
